requirements.txt
alembic==1.5.8
attrs==20.3.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
environs==9.3.2
Faker==8.0.0
Flask==1.1.2
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
gunicorn==20.1.0
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.11.1
marshmallow-sqlalchemy==0.24.2
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
PyJWT==2.0.1
pyparsing==2.4.7
pytest==6.2.3
python-dateutil==2.8.1
python-dotenv==0.17.0
python-editor==1.0.4
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.4.7
text-unidecode==1.3
toml==0.10.2
urllib3==1.25.9
Werkzeug==1.0.1
